requirements.txt
apprise==1.6.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
idna==3.4
loguru==0.7.2
Markdown==3.5.1
oauthlib==3.2.2
paho-mqtt==1.6.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.1.0
